requirements.txt
periodic-table-dataclasses 1.0
